agentRAG/app/rag/requirement.txt
fastapi==0.109.0
uvicorn==0.23.2
django==4.2.10
langchain==3.13.7
langchain-community==0.0.200
sentence-transformers==2.6.2
transformers==4.39.0
torch==2.2.0
pypdf==3.16.1
python-dotenv==1.1.1
requests==2.32.0